Refactor external AttachmentStore attachment logic to AttachmentSupport
Logic to upload attachment to AttachmentStore is same irrespective of ArtifactStore implementation. So move that to AttachementSupport trait for reuse
