+++ tests/src/test/scala/whisk/core/database/test/DbUtils.scala
@@ -296,7 +296,7 @@ trait DbUtils extends Assertions {
    */
   def inlinedAttachmentSize(db: ArtifactStore[_]): Int = {
     db match {
-      case inliner: AttachmentSupport =>
+      case inliner: AttachmentSupport[_] =>
         inliner.maxInlineSize.toBytes.toInt - 1
       case _ =>
         throw new IllegalStateException(s"ArtifactStore does not support attachment inlining $db")
@@ -308,7 +308,7 @@ trait DbUtils extends Assertions {
    */
   def nonInlinedAttachmentSize(db: ArtifactStore[_]): Int = {
     db match {
-      case inliner: AttachmentSupport =>
+      case inliner: AttachmentSupport[_] =>
         val inlineSize = inliner.maxInlineSize.toBytes.toInt
         val chunkSize = inliner.chunkSize.toBytes.toInt
         Math.max(inlineSize, chunkSize) * 2
